requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Jinja2==3.1.4
python-multipart==0.0.9
Pillow==10.4.0
PyMuPDF==1.24.10
